+++ without-algorithm/src/main/java/org/astar/vo/AStarPathFinder.java
@@ -1,4 +1,4 @@
-package org.astar.a;
+package org.astar.vo;
 
 
 import java.awt.Point;
